Add files via upload
This is another practice problem from "Automate the Boring Stuff With Python" by Al Sweigart, 2nd Edition, Page 108. I did a little bit more than required - I also calculated what the actual odds are of getting a given streak and compared that to the experimental value. For most streak sizes of ~4 - 11 there was about a 1-2% difference between the experimental and calculated values. Weird.